Seperate script for chart and table service.
