+++ varsel-frontend/src/pages/IndexPage.vue
@@ -28,25 +28,25 @@
             <div class="col-md-2 col-sm-6 col-xs-12">
               <q-card class="text-center">
                 <q-card-section>
-                  <div class="text-h6 text-green">{{ getMinPrice().toFixed(2) }} kr/kWh</div>
+                  <div class="text-h6 text-green">{{ getMinPrice(prices).toFixed(2) }} kr/kWh</div>
                   <div class="text-subtitle2">Laveste pris</div>
-                  <div class="text-caption">{{ getMinPriceTime() }}</div>
+                  <div class="text-caption">{{ getMinPriceTime(prices) }}</div>
                 </q-card-section>
               </q-card>
             </div>
             <div class="col-md-2 col-sm-6 col-xs-12">
               <q-card class="text-center">
                 <q-card-section>
-                  <div class="text-h6 text-red">{{ getMaxPrice().toFixed(2) }} kr/kWh</div>
+                  <div class="text-h6 text-red">{{ getMaxPrice(prices).toFixed(2) }} kr/kWh</div>
                   <div class="text-subtitle2">Høyeste pris</div>
-                  <div class="text-caption">{{ getMaxPriceTime() }}</div>
+                  <div class="text-caption">{{ getMaxPriceTime(prices) }}</div>
                 </q-card-section>
               </q-card>
             </div>
             <div class="col-md-2 col-sm-6 col-xs-12">
               <q-card class="text-center">
                 <q-card-section>
-                  <div class="text-h6 text-orange">{{ getAvgPrice().toFixed(2) }} kr/kWh</div>
+                  <div class="text-h6 text-orange">{{ getAvgPrice(prices).toFixed(2) }} kr/kWh</div>
                   <div class="text-subtitle2">Gjennomsnitt</div>
                   <div class="text-caption">{{ prices.length }} timer</div>
                 </q-card-section>
@@ -55,7 +55,7 @@
             <div class="col-md-2 col-sm-6 col-xs-12">
               <q-card class="text-center">
                 <q-card-section>
-                  <div class="text-h6 text-blue">{{ getPriceDifference().toFixed(2) }} kr/kWh</div>
+                  <div class="text-h6 text-blue">{{ getPriceDifference(prices).toFixed(2) }} kr/kWh</div>
                   <div class="text-subtitle2">Forskjell</div>
                   <div class="text-caption">Høyeste - Laveste</div>
                 </q-card-section>
@@ -74,80 +74,13 @@
 </template>
 
 <script setup lang="ts">
-import { ref, computed, watch, nextTick, onUnmounted } from 'vue'
+import { ref, watch, nextTick } from 'vue'
 import { api } from 'boot/axios'
-import type { QTableColumn } from 'quasar'
 import FooterSection from 'src/components/FooterSection.vue'
 import HeaderSection from 'src/components/HeaderSection.vue'
 import NavSection from 'src/components/NavSection.vue'
-import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  LineController,
-  Title,
-  Tooltip,
-  Legend,
-  Filler
-} from 'chart.js'
-
-// Register Chart.js components
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  LineController,
-  Title,
-  Tooltip,
-  Legend,
-  Filler
-)
-
-// Define options for area and city selection
-const areaOptions = [
-  { value: 'NO1', label: 'Øst-Norge' },
-  { value: 'NO2', label: 'Sør-Norge' },
-  { value: 'NO3', label: 'Midt-Norge' },
-  { value: 'NO4', label: 'Nord-Norge' },
-  { value: 'NO5', label: 'Vest-Norge' }
-]
-
-// Define specific cities with their areas
-const cityOptions = [
-  { value: '', label: 'Oslo', area: 'NO1' },
-  { value: 'Moss', label: 'Moss', area: 'NO1' },
-  { value: 'Drammen', label: 'Drammen', area: 'NO1' },
-  { value: 'Fredrikstad', label: 'Fredrikstad', area: 'NO1' },
-
-  { value: '', label: 'Kristiansand', area: 'NO2' },
-  { value: 'Arendal', label: 'Arendal', area: 'NO2' },
-  { value: 'Haugesund', label: 'Haugesund', area: 'NO2' },
-  { value: 'Porsgrunn', label: 'Porsgrunn', area: 'NO2' },
-  { value: 'Sandefjord', label: 'Sandefjord', area: 'NO2' },
-  { value: 'Stavanger', label: 'Stavanger', area: 'NO2' },
-  { value: 'Tonsberg', label: 'Tønsberg', area: 'NO2' },
-
-  { value: '', label: 'Trondheim', area: 'NO3' },
-  { value: 'Alesund', label: 'Ålesund', area: 'NO3' },
-
-  { value: '', label: 'Tromsø', area: 'NO4' },
-  { value: 'Bodo', label: 'Bodø', area: 'NO4' },
-
-  { value: '', label: 'Bergen', area: 'NO5' },
-]
-
-// Base cities for each area, used for default selection
-const baseCities: Record<'NO1' | 'NO2' | 'NO3' | 'NO4' | 'NO5', string> = {
-  NO1: 'Oslo',
-  NO2: 'Kristiansand',
-  NO3: 'Trondheim',
-  NO4: 'Tromsø',
-  NO5: 'Bergen'
-};
-
+import { useTableServices, type Price, baseCities } from 'src/components/Index/TableScript'
+import { useChartServices } from 'src/components/Index/ChartScript'
 
 // Reactive state variables
 const selectedArea = ref('NO1')
@@ -157,43 +90,22 @@ const startHour = ref<number | null>(null)
 const endHour = ref<number | null>(null)
 const prices = ref<Price[]>([])
 const isTaxIncluded = ref(false)
-const chartCanvas = ref<HTMLCanvasElement | null>(null)
-let chartInstance: ChartJS | null = null
 
+// Use table services
+const {
+  areaOptions,
+  columns,
+  filteredCityOptions,
+  getMinPrice,
+  getMaxPrice,
+  getAvgPrice,
+  getPriceDifference,
+  getMinPriceTime,
+  getMaxPriceTime
+} = useTableServices(selectedArea)
 
-
-const filteredCityOptions = computed(() =>
-  cityOptions.filter(city => city.area === selectedArea.value)
-)
-
-// Define the columns for the QTable
-const columns: QTableColumn[] = [
-  {
-    name: 'area',
-    label: 'Område',
-    field: row => areaOptions.find(option => option.value === row.area)?.label || row.area,
-    align: 'left' as const
-  },
-  { name: 'city', label: 'By', field: 'city', align: 'left' as const },
-  {
-    name: 'date', label: 'Dato', field: row => {
-      const [year, month, day] = row.date.split('-');
-      return `${day}.${month}.${year}`; // DD.MM.YYYY format
-    }, align: 'left' as const
-  },
-  { name: 'time_start', label: 'Start', field: row => row.time_start.slice(11, 16), align: 'left' as const }, // Only retrieves hour and minute
-  { name: 'time_end', label: 'Slutt', field: row => row.time_end.slice(11, 16), align: 'left' as const }, // Only retrieves hour and minute
-  { name: 'NOK_per_kWh', label: 'Pris (kr/kWh)', field: 'NOK_per_kWh', align: 'right' as const, format: (val: number) => `${val.toFixed(2)} kr` }
-]
-
-interface Price {
-  time_start: string;
-  time_end: string;
-  NOK_per_kWh: number;
-  area: string;
-  city: string;
-  date: string;
-}
+// Use chart services
+const { chartCanvas, createChart } = useChartServices()
 
 async function fetchPrices() {
   const storedTaxPreference = localStorage.getItem('isTaxIncluded');
@@ -237,7 +149,7 @@ async function fetchPrices() {
 
     // Create chart after data is loaded and DOM is updated
     await nextTick();
-    createChart();
+    createChart(prices.value);
   } catch (err) {
     console.error('Could not fetch prices', err);
   }
@@ -247,177 +159,4 @@ async function fetchPrices() {
 watch(selectedArea, () => {
   selectedCity.value = null;
 });
-
-// Clean up chart on component unmount
-onUnmounted(() => {
-  if (chartInstance) {
-    chartInstance.destroy();
-  }
-});
-
-// Helper functions for price statistics
-function getMinPrice(): number {
-  if (prices.value.length === 0) return 0;
-  return Math.min(...prices.value.map(p => p.NOK_per_kWh));
-}
-
-function getMaxPrice(): number {
-  if (prices.value.length === 0) return 0;
-  return Math.max(...prices.value.map(p => p.NOK_per_kWh));
-}
-
-function getAvgPrice(): number {
-  if (prices.value.length === 0) return 0;
-  const sum = prices.value.reduce((acc, p) => acc + p.NOK_per_kWh, 0);
-  return sum / prices.value.length;
-}
-
-function getPriceDifference(): number {
-  return getMaxPrice() - getMinPrice();
-}
-
-function getMinPriceTime(): string {
-  if (prices.value.length === 0) return '';
-  const minPrice = getMinPrice();
-  const minPriceData = prices.value.find(p => p.NOK_per_kWh === minPrice);
-  return minPriceData ? minPriceData.time_start.slice(11, 16) : '';
-}
-
-function getMaxPriceTime(): string {
-  if (prices.value.length === 0) return '';
-  const maxPrice = getMaxPrice();
-  const maxPriceData = prices.value.find(p => p.NOK_per_kWh === maxPrice);
-  return maxPriceData ? maxPriceData.time_start.slice(11, 16) : '';
-}
-
-
-// Function to create the price chart
-function createChart() {
-  if (!chartCanvas.value || prices.value.length === 0) {
-    return;
-  }
-
-  // Destroy existing chart if it exists
-  if (chartInstance) {
-    chartInstance.destroy();
-  }
-
-  const ctx = chartCanvas.value.getContext('2d');
-  if (!ctx) {
-    return;
-  }
-
-  const labels = prices.value.map(price => {
-    const timeStr = price.time_start || '';
-    return timeStr.length >= 16 ? timeStr.slice(11, 16) : timeStr;
-  });
-  const data = prices.value.map(price => typeof price.NOK_per_kWh === 'number' ? price.NOK_per_kWh : parseFloat(price.NOK_per_kWh) || 0);
-
-  // Calculate min, max, and average
-  const minPrice = Math.min(...data);
-  const maxPrice = Math.max(...data);
-  const avgPrice = data.reduce((sum, price) => sum + price, 0) / data.length;
-
-  // Find indices of min and max prices
-  const minIndex = data.indexOf(minPrice);
-  const maxIndex = data.indexOf(maxPrice);
-
-  try {
-    chartInstance = new ChartJS(ctx, {
-      type: 'line',
-      data: {
-        labels: labels,
-        datasets: [{
-          label: 'Pris (kr/kWh)',
-          data: data,
-          borderColor: 'rgb(75, 192, 192)',
-          backgroundColor: 'rgba(75, 192, 192, 0.2)',
-          tension: 0.1,
-          fill: true,
-          pointBackgroundColor: data.map((price, index) => {
-            if (index === minIndex) return 'green';
-            if (index === maxIndex) return 'red';
-            return 'rgb(75, 192, 192)';
-          }),
-          pointBorderColor: data.map((price, index) => {
-            if (index === minIndex) return 'darkgreen';
-            if (index === maxIndex) return 'darkred';
-            return 'rgb(75, 192, 192)';
-          }),
-          pointRadius: data.map((price, index) => {
-            if (index === minIndex || index === maxIndex) return 8;
-            return 4;
-          })
-        }]
-      },
-      options: {
-        responsive: true,
-        maintainAspectRatio: false,
-        plugins: {
-          title: {
-            display: true,
-            text: 'Strømpriser gjennom dagen'
-          },
-          legend: {
-            display: true
-          },
-          tooltip: {
-            callbacks: {
-              label: function(context) {
-                const value = context.parsed.y.toFixed(2);
-                let label = `Pris: ${value} kr/kWh`;
-
-                if (context.dataIndex === minIndex) {
-                  label += ' (Laveste)';
-                } else if (context.dataIndex === maxIndex) {
-                  label += ' (Høyeste)';
-                }
-
-                return label;
-              }
-            }
-          },
-          // @ts-expect-error - annotation plugin types may not be fully compatible
-          annotation: {
-            annotations: {
-              averageLine: {
-                type: 'line',
-                yMin: avgPrice,
-                yMax: avgPrice,
-                borderColor: 'orange',
-                borderWidth: 2,
-                borderDash: [5, 5],
-                label: {
-                  content: `Gjennomsnitt: ${avgPrice.toFixed(2)} kr/kWh`,
-                  enabled: true,
-                  position: 'end',
-                  backgroundColor: 'orange',
-                  color: 'white',
-                  padding: 4
-                }
-              }
-            }
-          }
-        },
-        scales: {
-          y: {
-            beginAtZero: false,
-            title: {
-              display: true,
-              text: 'Pris (kr/kWh)'
-            }
-          },
-          x: {
-            title: {
-              display: true,
-              text: 'Tid'
-            }
-          }
-        }
-      }
-    });
-  } catch (error) {
-    console.error('Error creating chart:', error);
-  }
-}
 </script>
